requirements.txt
supabase
python-dotenv
requests
google-generativeai
beautifulsoup4 